+++ src/mutators/index.ts
@@ -1,5 +1,6 @@
 // Export all mutators
 export * from "./array-mutators";
+export * from "./boolean-mutators";
 export * from "./date-mutators";
 export * from "./number-mutators";
 export * from "./object-mutators";